applied more review suggestions
