+++ ActivatePubKey/__tests__/handler.test.ts
@@ -4,6 +4,7 @@ import * as E from "fp-ts/lib/Either";
 import {
   AssertionFileName,
   LolliPOPKeysModel,
+  NewLolliPopPubKeys,
   RetrievedLolliPopPubKeys,
   TTL_VALUE_AFTER_UPDATE
 } from "../../model/lollipop_keys";
@@ -12,9 +13,14 @@ import { AssertionTypeEnum } from "@pagopa/io-functions-commons/dist/generated/d
 import { PubKeyStatusEnum } from "../../generated/definitions/internal/PubKeyStatus";
 import { NonNegativeInteger } from "@pagopa/ts-commons/lib/numbers";
 import { ActivatePubKeyHandler } from "../handler";
-import { BlobService, createBlobService } from "azure-storage";
-import { getPopDocumentReader } from "../../utils/readers";
-import { getAssertionWriter, getPopDocumentWriter } from "../../utils/writers";
+import { BlobService } from "azure-storage";
+import { getPopDocumentReader, PopDocumentReader } from "../../utils/readers";
+import {
+  AssertionWriter,
+  getAssertionWriter,
+  getPopDocumentWriter,
+  PopDocumentWriter
+} from "../../utils/writers";
 import { ActivatePubKeyPayload } from "../../generated/definitions/internal/ActivatePubKeyPayload";
 import {
   retrievedLollipopKeysToApiActivatedPubKey,
@@ -24,12 +30,17 @@ import * as fn_commons from "@pagopa/io-functions-commons/dist/src/utils/azure_s
 import { getAllAssertionsRef } from "../../utils/lollipopKeys";
 import { JwkPubKeyHashAlgorithmEnum } from "../../generated/definitions/internal/JwkPubKeyHashAlgorithm";
 import {
+  aRetrievedPendingLollipopPubKeySha256,
+  aRetrievedValidLollipopPubKeySha256,
   aValidJwk,
   aValidSha256AssertionRef,
   aValidSha512AssertionRef,
   toEncodedJwk
 } from "../../__mocks__/lollipopPubKey.mock";
 import { blobServiceMock } from "../../__mocks__/blobService.mock";
+import { AssertionRef } from "../../generated/definitions/internal/AssertionRef";
+import { ErrorKind } from "../../utils/errors";
+import { contextMock } from "../../__mocks__/context.mock";
 
 const aFiscalCode = "SPNDNL80A13Y555X" as FiscalCode;
 
@@ -82,9 +93,40 @@ const lollipopPubKeysModelMock = ({
   upsert: upsertMock
 } as unknown) as LolliPOPKeysModel;
 
-const contextMock = {} as any;
-
 const LOLLIPOP_ASSERTION_STORAGE_CONTAINER_NAME = "assertions" as NonEmptyString;
+
+const popDocumentReaderMock = jest.fn(
+  (assertionRef: AssertionRef) =>
+    TE.of({
+      ...aRetrievedPendingLollipopPubKeySha256,
+      assertionRef: assertionRef,
+      id: `${assertionRef}-000000`,
+      version: 0
+    }) as ReturnType<PopDocumentReader>
+);
+
+const popDocumentWriterMock = jest.fn(
+  (item: NewLolliPopPubKeys) =>
+    TE.of({
+      ...aRetrievedPendingLollipopPubKeySha256,
+      ...item,
+      id: `${item.assertionRef}-000001`,
+      version: 1
+    }) as ReturnType<PopDocumentWriter>
+);
+const assertionWriterMock = jest.fn(
+  () => TE.of(true) as ReturnType<AssertionWriter>
+);
+
+var expiresAtDate = new Date(); // Now
+expiresAtDate.setDate(expiresAtDate.getDate() + 30); // Set now + 30 days as the new date
+
+const aValidPayload: ActivatePubKeyPayload = {
+  fiscal_code: aFiscalCode,
+  assertion: "an assertion" as NonEmptyString,
+  assertion_type: AssertionTypeEnum.SAML,
+  expires_at: expiresAtDate
+};
 
 describe("activatePubKey handler", () => {
   beforeEach(() => {
@@ -256,5 +298,228 @@ describe("activatePubKey handler", () => {
         aValidRetrievedPopDocumentWithMasterAlgo
       )
     });
+  });
+});
+
+describe("ActivatePubKey - Errors", () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it("should return 404 NotFound when assertionRef doen not exists", async () => {
+    popDocumentReaderMock.mockImplementationOnce(() =>
+      TE.left({ kind: ErrorKind.NotFound })
+    );
+
+    const handler = ActivatePubKeyHandler(
+      popDocumentReaderMock,
+      popDocumentWriterMock,
+      assertionWriterMock
+    );
+
+    const res = await handler(
+      contextMock,
+      aValidSha256AssertionRef,
+      aValidPayload
+    );
+
+    expect(res).toMatchObject({
+      kind: "IResponseErrorNotFound",
+      detail: "NotFound: Could not find requested resource"
+    });
+
+    expect(popDocumentReaderMock).toHaveBeenCalledWith(
+      aValidSha256AssertionRef
+    );
+    expect(assertionWriterMock).not.toHaveBeenCalled();
+    expect(popDocumentWriterMock).not.toHaveBeenCalled();
+  });
+
+  it("should return 500 Internal Error when a pop document with status != PENDING is found", async () => {
+    popDocumentReaderMock.mockImplementationOnce(assertionRef =>
+      TE.of({
+        ...aRetrievedValidLollipopPubKeySha256,
+        assertionRef: assertionRef,
+        id: `${assertionRef}-000000` as NonEmptyString,
+        version: 0 as NonNegativeInteger,
+        status: PubKeyStatusEnum.REVOKED
+      })
+    );
+
+    const handler = ActivatePubKeyHandler(
+      popDocumentReaderMock,
+      popDocumentWriterMock,
+      assertionWriterMock
+    );
+
+    const res = await handler(
+      contextMock,
+      aValidSha256AssertionRef,
+      aValidPayload
+    );
+
+    expect(res).toMatchObject({
+      kind: "IResponseErrorInternal",
+      detail:
+        "Internal server error: Unexpected status on pop document during activation: REVOKED"
+    });
+
+    expect(popDocumentReaderMock).toHaveBeenCalledWith(
+      aValidSha256AssertionRef
+    );
+    expect(assertionWriterMock).not.toHaveBeenCalled();
+    expect(popDocumentWriterMock).not.toHaveBeenCalled();
+  });
+
+  it("should return 500 Internal Error when an error occurred reading document", async () => {
+    popDocumentReaderMock.mockImplementationOnce(() =>
+      TE.left({ kind: ErrorKind.Internal, detail: "an Error" })
+    );
+
+    const handler = ActivatePubKeyHandler(
+      popDocumentReaderMock,
+      popDocumentWriterMock,
+      assertionWriterMock
+    );
+
+    const res = await handler(
+      contextMock,
+      aValidSha256AssertionRef,
+      aValidPayload
+    );
+
+    expect(res).toMatchObject({
+      kind: "IResponseErrorInternal",
+      detail: "Internal server error: an Error"
+    });
+
+    expect(popDocumentReaderMock).toHaveBeenCalledWith(
+      aValidSha256AssertionRef
+    );
+    expect(assertionWriterMock).not.toHaveBeenCalled();
+    expect(popDocumentWriterMock).not.toHaveBeenCalled();
+  });
+
+  it("should return 500 Internal Error when an error occurred writing assertion into storage", async () => {
+    assertionWriterMock.mockImplementationOnce(() =>
+      TE.left({ kind: ErrorKind.Internal, detail: "an Error on storage" })
+    );
+
+    const handler = ActivatePubKeyHandler(
+      popDocumentReaderMock,
+      popDocumentWriterMock,
+      assertionWriterMock
+    );
+
+    const res = await handler(
+      contextMock,
+      aValidSha256AssertionRef,
+      aValidPayload
+    );
+
+    expect(res).toMatchObject({
+      kind: "IResponseErrorInternal",
+      detail: "Internal server error: an Error on storage"
+    });
+
+    const expectedResult = {
+      assertionFileName: `${aFiscalCode}-${aValidSha256AssertionRef}`
+    };
+
+    expect(popDocumentReaderMock).toHaveBeenCalledWith(
+      aValidSha256AssertionRef
+    );
+    expect(assertionWriterMock).toHaveBeenCalledWith(
+      expectedResult.assertionFileName,
+      aValidPayload.assertion
+    );
+    expect(popDocumentWriterMock).not.toHaveBeenCalled();
+  });
+
+  it("should return 500 Internal Error when an error occurred storing master key", async () => {
+    popDocumentWriterMock.mockImplementationOnce(() =>
+      TE.left({ kind: ErrorKind.Internal, detail: "an Error on cosmos update" })
+    );
+
+    const handler = ActivatePubKeyHandler(
+      popDocumentReaderMock,
+      popDocumentWriterMock,
+      assertionWriterMock
+    );
+
+    const res = await handler(
+      contextMock,
+      aValidSha256AssertionRef,
+      aValidPayload
+    );
+
+    expect(res).toMatchObject({
+      kind: "IResponseErrorInternal",
+      detail: "Internal server error: an Error on cosmos update"
+    });
+
+    const expectedResult = {
+      assertionFileName: `${aFiscalCode}-${aValidSha256AssertionRef}`
+    };
+
+    expect(popDocumentReaderMock).toHaveBeenCalledWith(
+      aValidSha256AssertionRef
+    );
+    expect(assertionWriterMock).toHaveBeenCalledWith(
+      expectedResult.assertionFileName,
+      aValidPayload.assertion
+    );
+    expect(popDocumentWriterMock).toHaveBeenCalledTimes(1);
+  });
+
+  it("should return 500 Internal Error when an error occurred storing used key", async () => {
+    popDocumentWriterMock
+      // First insert OK
+      .mockImplementationOnce(
+        (item: NewLolliPopPubKeys) =>
+          TE.of({
+            ...aRetrievedPendingLollipopPubKeySha256,
+            ...item,
+            id: `${item.assertionRef}-000001`,
+            version: 1
+          }) as ReturnType<PopDocumentWriter>
+      )
+      // Second insert KO
+      .mockImplementationOnce(() =>
+        TE.left({
+          kind: ErrorKind.Internal,
+          detail: "an Error on cosmos update"
+        })
+      );
+
+    const handler = ActivatePubKeyHandler(
+      popDocumentReaderMock,
+      popDocumentWriterMock,
+      assertionWriterMock
+    );
+
+    const res = await handler(
+      contextMock,
+      aValidSha256AssertionRef,
+      aValidPayload
+    );
+
+    expect(res).toMatchObject({
+      kind: "IResponseErrorInternal",
+      detail: "Internal server error: an Error on cosmos update"
+    });
+
+    const expectedResult = {
+      assertionFileName: `${aFiscalCode}-${aValidSha256AssertionRef}`
+    };
+
+    expect(popDocumentReaderMock).toHaveBeenCalledWith(
+      aValidSha256AssertionRef
+    );
+    expect(assertionWriterMock).toHaveBeenCalledWith(
+      expectedResult.assertionFileName,
+      aValidPayload.assertion
+    );
+    expect(popDocumentWriterMock).toHaveBeenCalledTimes(2);
   });
 });
